BAMPFA-582: moved utils.js one directory up
